+++ lib/src/slider_clipper.dart
@@ -1,8 +1,12 @@
 part of 'image_compare_slider.dart';
 
 class _SliderClipper extends CustomClipper<Rect> {
-  _SliderClipper({required this.position, required this.direction});
+  _SliderClipper({required this.position, required this.direction})
+      : invert = false;
+  _SliderClipper.invert({required this.position, required this.direction})
+      : invert = true;
 
+  final bool invert;
   final double position;
   final SliderDirection direction;
 
@@ -10,17 +14,37 @@ class _SliderClipper extends CustomClipper<Rect> {
   Rect getClip(Size size) {
     switch (direction) {
       case SliderDirection.leftToRight:
-        final dx = position * size.width;
-        return Rect.fromLTRB(0, 0, dx, size.height);
+        if (invert) {
+          final dx = position * size.width;
+          return Rect.fromLTRB(0, 0, dx, size.height);
+        } else {
+          final dx = position * size.width;
+          return Rect.fromLTRB(dx, 0, size.width, size.height);
+        }
       case SliderDirection.rightToLeft:
-        final dx = position * size.width;
-        return Rect.fromLTRB(dx, 0, size.width, size.height);
+        if (invert) {
+          final dx = position * size.width;
+          return Rect.fromLTRB(dx, 0, size.width, size.height);
+        } else {
+          final dx = position * size.width;
+          return Rect.fromLTRB(0, 0, dx, size.height);
+        }
       case SliderDirection.topToBottom:
-        final dy = position * size.height;
-        return Rect.fromLTRB(0, 0, size.width, dy);
+        if (invert) {
+          final dy = position * size.height;
+          return Rect.fromLTRB(0, 0, size.width, dy);
+        } else {
+          final dy = position * size.height;
+          return Rect.fromLTRB(0, dy, size.width, size.height);
+        }
       case SliderDirection.bottomToTop:
-        final dy = position * size.height;
-        return Rect.fromLTRB(0, dy, size.width, size.height);
+        if (invert) {
+          final dy = position * size.height;
+          return Rect.fromLTRB(0, dy, size.width, size.height);
+        } else {
+          final dy = position * size.height;
+          return Rect.fromLTRB(0, 0, size.width, dy);
+        }
     }
   }